feat: Add all screenshots
