--- src/Projects/Projects.css
@@ -1,1 +1,0 @@
-/* Projects.css */
