+++ src/components/Intro/Intro.css
@@ -17,7 +17,7 @@
 
 .intro__container {
   background-color: white;
-  height: 20%;
+  height: 30%;
   width: 45%;
   max-width: 500px;
   display: flex;
