+++ src/components/Projects/Projects.css
@@ -1,0 +1,36 @@
+/* Projects.css */
+
+.projects {
+  width: 100vw;
+  height: 100vh;
+  margin: 0;
+}
+
+.projects h1 {
+  margin: 2rem 1rem;
+}
+
+.projects__container {
+  max-width: 100vw;
+  display: flex;
+  overflow-x: scroll;
+  justify-content: center;
+  align-items: center;
+}
+
+.projects__item {
+  margin: 1rem;
+  width: 350px;
+  border-radius: 10px;
+  overflow: hidden;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+
+}
+
+.projects__item-image {
+  width: 100%;
+  height: 100%;
+  border-radius: 10px;
+  object-fit: cover;
+  transition: transform 0.3s ease-in-out;
+}
